+++ changelog.md
@@ -3,6 +3,7 @@
 v1.15.0:
 
 - Support The War Within/11.0
+- Add support for exporting and importing profiles through text.
 
 v1.14.42: